Move helper function closer to usage
